requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-socketio==5.3.7
google-generativeai==0.8.3
sentence-transformers==3.3.1
playwright==1.49.0
apscheduler==3.10.4
python-dotenv==1.0.1
eventlet==0.37.0
cryptography==43.0.0
